move all schemas into dedicated directory
