+++ src/lib/memberIdBundle/MemberIdBundle.ts
@@ -2,14 +2,13 @@ import { type TrustAnchor } from '@relaycorp/dnssec';
 import { type Certificate as CertificateSchema } from '@peculiar/asn1-x509';
 import { AsnSerializer } from '@peculiar/asn1-schema';
 
-import { type DnssecChainSchema } from '../dns/DnssecChainSchema.js';
+import { type DnssecChainSchema } from '../schemas/DnssecChainSchema.js';
 import { type DatePeriod } from '../utils/DatePeriod.js';
 import Certificate from '../utils/x509/Certificate.js';
 import VeraError from '../VeraError.js';
 import { VeraDnssecChain } from '../dns/VeraDnssecChain.js';
 import { getKeySpec } from '../dns/organisationKeys.js';
-
-import { MemberIdBundleSchema } from './MemberIdBundleSchema.js';
+import { MemberIdBundleSchema } from '../schemas/MemberIdBundleSchema.js';
 
 async function verifyCertificateChain(
   orgCertificate: Certificate,
