+++ src/lib/schemas/VeraSignatureSchema.ts
@@ -1,19 +1,15 @@
 /* eslint-disable new-cap */
 
-import { AsnProp, AsnPropTypes } from '@peculiar/asn1-schema';
+import { AsnProp } from '@peculiar/asn1-schema';
 import { SignedData } from '@peculiar/asn1-cms';
 import { Certificate } from '@peculiar/asn1-x509';
 
-import { DnssecChainSchema } from '../dns/DnssecChainSchema.js';
-
-import { DatePeriodSchema } from './DatePeriodSchema.js';
+import { DnssecChainSchema } from './DnssecChainSchema.js';
+import { VeraSignatureMetadataSchema } from './VeraSignatureMetadataSchema.js';
 
 export class VeraSignatureSchema {
-  @AsnProp({ type: DatePeriodSchema })
-  public validityPeriod!: DatePeriodSchema;
-
-  @AsnProp({ type: AsnPropTypes.VisibleString })
-  public serviceOid!: AsnPropTypes.VisibleString;
+  @AsnProp({ type: VeraSignatureMetadataSchema })
+  public metadata!: VeraSignatureMetadataSchema;
 
   @AsnProp({ type: DnssecChainSchema })
   public dnssecChain!: DnssecChainSchema;
